IntelliJ Git Test, static removal & improvements to Cache Downloader.

+++ RSPS Launcher/src/main/java/com/tutus/download/cache/CacheDownloader.java
@@ -5,26 +5,30 @@ import javafx.scene.control.Alert;
 import javafx.scene.control.ButtonType;
 import net.lingala.zip4j.ZipFile;
 import net.lingala.zip4j.exception.ZipException;
+import net.lingala.zip4j.progress.ProgressMonitor;
 
 import java.io.*;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
 
 public class CacheDownloader {
 
+    private boolean cacheUpdated = false;
+
 // downloads the Cache File
-    public static void downloadCache() throws IOException
+    public void downloadCache() throws IOException
     {
         //deletes jar before downloading new one (if it exists)
         File file = new File(Configuration.CACHE_SAVE_DIR + Configuration.CACHE_SAVE_NAME);
-        file.delete();
+        if (file.exists()) {
+            file.delete();
+        }
         System.out.println("Deleted File");
 
         // Downloads the cache file
         InputStream inputStream = new URL(Configuration.CACHE_DOWNLOAD_URL).openStream();
-        Files.copy(inputStream, Paths.get(Configuration.CACHE_SAVE_DIR + Configuration.CACHE_SAVE_NAME), StandardCopyOption.REPLACE_EXISTING);
+        Files.copy(inputStream, Paths.get(Configuration.CACHE_SAVE_DIR + Configuration.CACHE_SAVE_NAME));
         System.out.println("Downloaded File");
 
         // Get this download's progress.
@@ -33,37 +37,40 @@ public class CacheDownloader {
         // Unzips the Cache's Content
         unzipCache();
         // Sends a notification that the Cache has been updated (TEMPORARILY)
-        cacheUpdated();
+
+        if(cacheUpdated) {
+            Alert downloadFinished = new Alert(Alert.AlertType.INFORMATION);
+            downloadFinished.setTitle("Cache Update Completed");
+            downloadFinished.setHeaderText("Your Cache has been updated completely.");
+            System.out.println("Alert Shown");
+            downloadFinished.showAndWait().ifPresent(rs -> {
+                if (rs == ButtonType.OK) {
+                    System.out.println("Pressed OK.");
+
+                }
+
+            });
+            //cacheUpdated(); // Use a boolean for this. Don't rely on your exceptions to be thrown
+        }
     }
 
     private static String zipFilePath = Configuration.CACHE_SAVE_DIR + Configuration.CACHE_SAVE_NAME;
     private static String destDir = Configuration.CACHE_SAVE_DIR + Configuration.CACHE_FOLDER_NAME;
 
-    private static void unzipCache() throws IOException {
+    private void unzipCache() throws IOException {
         System.out.println("Unzipping the Cache");
         ZipFile zipFile = new ZipFile(zipFilePath);
         try {
-            //ZipFile zipFile = new ZipFile(zipFilePath);
             zipFile.extractAll(destDir);
+            if (zipFile.getProgressMonitor().getResult() == ProgressMonitor.Result.SUCCESS) {
+                cacheUpdated = true;
+            }
         } catch (ZipException e) {
             e.printStackTrace();
-        }
-        finally {
+        } finally {
             System.out.println("Removing the old zip file");
             Files.deleteIfExists(Paths.get(zipFilePath));
+            cacheUpdated = true;
         }
-    }
-
-// Sends an alert upon completion
-    public static void cacheUpdated() throws ZipException {
-        Alert downloadFinished = new Alert(Alert.AlertType.INFORMATION);
-        downloadFinished.setTitle("Cache Update Completed");
-        downloadFinished.setHeaderText("Your Cache has been updated completely.");
-        System.out.println("Alert Shown");
-        downloadFinished.showAndWait().ifPresent(rs -> {
-            if (rs == ButtonType.OK) {
-                System.out.println("Pressed OK.");
-            }
-        });
     }
 }
